move navbar handler to NavBar component
