+++ src/components/Header/Header.tsx
@@ -6,30 +6,8 @@ import { Navbar } from '../Navbar/Navbar'
 import { SearchBar } from './SearchBar'
 
 export const Header = () => {
-  function toggleNavbar() {
-    const el = document.getElementById('nav-menu');
-    var pos = 0;
-    var id = setInterval(frame, 5);
-    function frame() {
-      if (pos == 600) {
-        clearInterval(id);
-      } else {
-        pos += 10
-        if (el) {
-          el.style.height = pos + 'px';
-        }
-      }
-    }
-    const navbarEl = document.getElementById('nav-bar')
-    if (navbarEl) {
-      if (navbarEl.style.display === 'none') {
-        navbarEl.style.display = 'block'
-      } else {
-        navbarEl.style.display = 'none'
-      }
-    }
-  }
   const [showSearchBar, setShowSearchBar] = React.useState(false)
+  const [showNavBar, setShowNavBar] = React.useState(false)
 
   return (
     <React.Fragment>
@@ -39,7 +17,7 @@ export const Header = () => {
             icon={faBars}
             size="2x"
             cursor="pointer"
-            onClick={toggleNavbar}
+            onClick={() => setShowNavBar(!showNavBar)}
           />
           <span style={{ fontSize: '1.6rem', fontWeight: 'bold' }}>
             Squard
@@ -59,7 +37,7 @@ export const Header = () => {
       </div>
       <SearchBar show={showSearchBar} />
       <div id="nav-bar" style={{ display: 'none' }} className={styles.navModal}>
-        <Navbar />
+        <Navbar show={showNavBar} hideNavBar={() => setShowNavBar(false)}/>
       </div>
     </React.Fragment>
   )
